extract login logic to useLogin hook
