+++ src/Apps/erp/shared/auth/Login.tsx
@@ -1,5 +1,4 @@
-import React, { useCallback, useEffect, useReducer, useState } from 'react';
-import { useForm } from 'react-hook-form';
+import React from 'react';
 import Box from '@mui/material/Box';
 import TextField from '@mui/material/TextField';
 import Button from '@mui/material/Button';
@@ -13,6 +12,7 @@ import Visibility from '@mui/icons-material/Visibility';
 import VisibilityOff from '@mui/icons-material/VisibilityOff';
 
 import './Login.less';
+import { useLogin } from './useLogin';
 
 interface LoginProps {
   onLogin: (credentials: { username: string; password: string }) => void;
@@ -20,100 +20,17 @@ interface LoginProps {
   clientLogoUrl: string;
 }
 
-interface SessionState {
-  username: string;
-  roles: string[];
-  token: string;
-}
-
-type SessionAction = { type: 'SET'; payload: SessionState } | { type: 'CLEAR' };
-
-const sessionReducer = (state: SessionState, action: SessionAction): SessionState => {
-  switch (action.type) {
-    case 'SET':
-      return action.payload;
-    case 'CLEAR':
-      return { username: '', roles: [], token: '' };
-    default:
-      return state;
-  }
-};
-
 export const Login: React.FC<LoginProps> = ({ onLogin, onRecoverPassword, clientLogoUrl }) => {
-  const [session, dispatch] = useReducer(sessionReducer, {
-    username: '',
-    roles: [],
-    token: ''
-  });
-  const [showPassword, setShowPassword] = useState(false);
-
   const {
     register,
     handleSubmit,
-    setValue,
-    watch,
-    formState: { errors }
-  } = useForm<{ username: string; password: string; remember: boolean }>({
-    defaultValues: { username: '', password: '', remember: false }
-  });
-
-  useEffect(() => {
-    const saved = localStorage.getItem('login_username');
-    if (saved) {
-      setValue('username', saved);
-      setValue('remember', true);
-    }
-  }, [setValue]);
-
-  const remember = watch('remember');
-
-  const toggleShowPassword = useCallback(() => {
-    setShowPassword((prev) => !prev);
-  }, []);
-
-  const onSubmit = useCallback(
-    (data: { username: string; password: string; remember: boolean }) => {
-      if (data.remember) {
-        localStorage.setItem('login_username', data.username);
-      } else {
-        localStorage.removeItem('login_username');
-      }
-      onLogin({ username: data.username, password: data.password });
-      dispatch({
-        type: 'SET',
-        payload: {
-          username: data.username,
-          roles: ['user'],
-          token: 'mock-token'
-        }
-      });
-    },
-    [onLogin]
-  );
-
-  const handleRecover = useCallback(
-    (e?: React.MouseEvent<HTMLAnchorElement>) => {
-      if (e) e.preventDefault();
-      onRecoverPassword();
-    },
-    [onRecoverPassword]
-  );
-
-  useEffect(() => {
-    const handler = (e: KeyboardEvent) => {
-      if (e.ctrlKey && e.key === 'Enter') {
-        handleSubmit(onSubmit)();
-      }
-      if (e.altKey && e.key.toLowerCase() === 'r') {
-        handleRecover();
-      }
-      if (e.altKey && e.key.toLowerCase() === 'v') {
-        toggleShowPassword();
-      }
-    };
-    window.addEventListener('keydown', handler);
-    return () => window.removeEventListener('keydown', handler);
-  }, [handleSubmit, onSubmit, handleRecover, toggleShowPassword]);
+    errors,
+    onSubmit,
+    toggleShowPassword,
+    handleRecover,
+    showPassword,
+    session
+  } = useLogin({ onLogin, onRecoverPassword });
 
   return (
     <Box className="login-container">
